+++ packages/core/src/utils/clsx.ts
@@ -27,6 +27,7 @@ function toVal(mix: any) {
       const len = mix.length;
       for (k = 0; k < len; k++) {
         if (mix[k] && (y = toVal(mix[k]))) {
+          // oxlint-disable-next-line no-unused-expressions
           str && (str += ' ');
           str += y;
         }
@@ -34,6 +35,7 @@ function toVal(mix: any) {
     } else {
       for (y in mix) {
         if (mix[y]) {
+          // oxlint-disable-next-line no-unused-expressions
           str && (str += ' ');
           str += y;
         }
@@ -58,6 +60,7 @@ export function clsx(...args: ClassValue[]): string {
   const len = args.length;
   for (; i < len; i++) {
     if ((tmp = args[i]) && (x = toVal(tmp))) {
+      // oxlint-disable-next-line no-unused-expressions
       str && (str += ' ');
       str += x;
     }