+++ packages/core/src/utils/formatQuery/__tests__/dbquery.prisma.postgres.test.ts
@@ -1,6 +1,5 @@
 import { PGlite } from '@electric-sql/pglite';
 import { PrismaClient } from '@prisma/client';
-import type { SqlDriverAdapterFactory } from '@prisma/client/runtime/library';
 import { PrismaPGlite } from 'pglite-prisma-adapter';
 import type { TestSQLParams } from '../dbqueryTestUtils';
 import { CREATE_INDEX, CREATE_TABLE, dbTests, fields, superUsers } from '../dbqueryTestUtils';
@@ -10,7 +9,9 @@ const db = new PGlite();
 
 const superUsersPostgres = superUsers('postgres');
 
-const adapter = new PrismaPGlite(db) as SqlDriverAdapterFactory;
+const adapter = new PrismaPGlite(db);
+// oxlint-disable-next-line ban-ts-comment
+// @ts-ignore This only fails after generating the adapter
 const prisma = new PrismaClient({ adapter });
 
 beforeAll(async () => {