+++ packages/core/src/utils/formatQuery/defaultRuleGroupProcessorJSONata.ts
@@ -24,39 +24,67 @@ export const defaultRuleGroupProcessorJSONata: RuleGroupProcessor<string> = (
     validationMap,
   } = options;
 
-  const processRuleGroup = (rg: RuleGroupTypeAny, outermost?: boolean) => {
+  const processRuleGroup = (rg: RuleGroupTypeAny, outermost?: boolean): string => {
     if (!isRuleOrGroupValid(rg, validationMap[rg.id ?? /* istanbul ignore next */ ''])) {
       return outermost ? fallbackExpression : '';
     }
 
-    const expression: string = rg.rules
-      .map(rule => {
-        if (typeof rule === 'string') {
-          return rule;
+    const processedRules = [];
+    let precedingCombinator = '';
+    let firstRule = true;
+
+    for (const rule of rg.rules) {
+      // Independent combinators
+      if (typeof rule === 'string') {
+        precedingCombinator = rule;
+        continue;
+      }
+
+      // Groups
+      if (isRuleGroup(rule)) {
+        const processedGroup = processRuleGroup(rule);
+        if (processedGroup) {
+          if (!firstRule && precedingCombinator) {
+            processedRules.push(precedingCombinator);
+            precedingCombinator = '';
+          }
+          firstRule = false;
+          processedRules.push(processedGroup);
         }
-        if (isRuleGroup(rule)) {
-          return processRuleGroup(rule);
+        continue;
+      }
+
+      // Rules
+      const [validationResult, fieldValidator] = validateRule(rule);
+      if (
+        !isRuleOrGroupValid(rule, validationResult, fieldValidator) ||
+        rule.field === placeholderFieldName ||
+        rule.operator === placeholderOperatorName ||
+        /* istanbul ignore next */
+        (placeholderValueName !== undefined && rule.value === placeholderValueName)
+      ) {
+        continue;
+      }
+
+      const fieldData = getOption(fields, rule.field);
+      const processedRule = ruleProcessor(rule, {
+        ...options,
+        parseNumbers: getParseNumberBoolean(fieldData?.inputType),
+        escapeQuotes: (rule.valueSource ?? 'value') === 'value',
+        fieldData,
+      });
+
+      if (processedRule) {
+        if (!firstRule && precedingCombinator) {
+          processedRules.push(precedingCombinator);
+          precedingCombinator = '';
         }
-        const [validationResult, fieldValidator] = validateRule(rule);
-        if (
-          !isRuleOrGroupValid(rule, validationResult, fieldValidator) ||
-          rule.field === placeholderFieldName ||
-          rule.operator === placeholderOperatorName ||
-          /* istanbul ignore next */
-          (placeholderValueName !== undefined && rule.value === placeholderValueName)
-        ) {
-          return '';
-        }
-        const fieldData = getOption(fields, rule.field);
-        return ruleProcessor(rule, {
-          ...options,
-          parseNumbers: getParseNumberBoolean(fieldData?.inputType),
-          escapeQuotes: (rule.valueSource ?? 'value') === 'value',
-          fieldData,
-        });
-      })
-      .filter(Boolean)
-      .join(isRuleGroupType(rg) ? ` ${rg.combinator} ` : ' ');
+        firstRule = false;
+        processedRules.push(processedRule);
+      }
+    }
+
+    const expression = processedRules.join(isRuleGroupType(rg) ? ` ${rg.combinator} ` : ' ');
 
     const [prefix, suffix] = rg.not || !outermost ? [`${rg.not ? '$not' : ''}(`, ')'] : ['', ''];