+++ packages/rules-engine/src/hooks/useMergeComponents.ts
@@ -4,6 +4,7 @@ import type { ComponentsRE } from '../types';
 
 export const useMergeComponents = (componentsProp: Partial<ComponentsRE>): ComponentsRE =>
   useMemo(
+    // TODO: more granular merge strategy
     (): ComponentsRE => Object.assign({}, defaultComponentsRE, componentsProp),
     [componentsProp]
   );